solucion a los conflictos

+++ js/adminfilm.js
@@ -85,10 +85,12 @@ const deleteContent = async (film) => {
   let content = await result.json ()
   let idLine = content.find (para => para.id == film.id)
   let id = idLine.id
-  
-  await fetch (`http://localhost:3000/films/${id}`, {
-    method: 'DELETE',
-  });
+  let confirmThis = confirm ('¿Está seguro de que desea eliminar este item?')
+  if (confirmThis === true) {
+    await fetch (`http://localhost:3000/films/${id}`, {
+      method: 'DELETE',
+    });
+  }
 
 };
 
@@ -223,7 +225,7 @@ const spotLight = async (film) => {
         'Content-type': 'application/json; charset=UTF-8',
       }
     }) 
-    alert ('uwu')   
+    alert ('Contenido destacado con éxito')   
   }
 };
